Only search the lib dir for rpc_backend
The grep in rpc_backend uses . to search all files. Unfortunately
the current directory when the command is called is not the lib
directory, so change it to search the lib directory instead.
Without this fix the command is forced to grep all files under
devstack which could be very slow if one has downloaded large
images.

Change-Id: Ic8a59885cd07e5d6f2e734e80824762f9ac12996