Update patch set 1
Patch Set 1:

Chmouel, I don't mind the patch specifically, but I want to be sure we set appropriate expectations about support for shells other than bash.  The docs and HACKING.rst are clear that bash v3+ is what is supported so this is likely  to come up again.

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0