Additional sanity checks for RHEL6
Simple check to see if EPEL repo is enabled for RHEL6.  Installing
repos such as EPEL may be a personal preference, so we give the user
the URL to fix it rather than any automated action.

Also adds an additional sanity check which checking for pip.  In the
prior case of EPEL being disabled, a RHEL6 host may not have access to
python-pip.  Although this shouldn't happen, its a good sanity check.

Change-Id: I4e8a4dda9475b75d071534d4eef469198502a048