Install EPEL for rhel6 if not detected
Simple check to see if EPEL repo is enabled for RHEL6 and, if not,
automated install of the repo RPM.

Also adds an additional sanity check which checking for pip.  In the
prior case of EPEL being disabled, a RHEL6 host may not have access to
python-pip.  Although this shouldn't happen, its a good sanity check.

Change-Id: I4e8a4dda9475b75d071534d4eef469198502a048