Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

If enabling EPEL is not to be automatic, I would suggest that it be possible for devstack to enable EPEL if explicitly configured to do so (i.e. ENABLE_EPEL=True).  The fact that it is a personal preference shouldn't require manual intervention be for devstack to be able to work, and it would reflect poorly on RHEL if the process for running devstack was more involved than for other distros.

Patch-set: 1
Label: Code-Review=-1