Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

Now that EPEL is to be assured of being enabled, is python-pip guaranteed to be installed?  I see that python-pip is in files/rpms/general already, and I'm a bit confused as to why devstack wasn't previously failing if it couldn't install it.  If package installation failure will halt devstack, as I would assume it should, the check for pip should be redundant.

Patch-set: 3
Label: Code-Review=-1