Update patch set 3
Patch Set 3:

> and I'm a bit confused as to why devstack wasn't previously failing if it couldn't install it. 

> If package installation failure will halt devstack, as I would assume it should, the check for pip should be redundant.

This is a separate bug that I'm thinking about.  We use "yum --assumeyes" to install without the "is this ok" prompt, but it seems that that also assumes that you're ok with missing packages.

I'm trying to figure out if there's a "yes, assume i say yes to installing those packages but stop if any are missing" flag

Patch-set: 3