Update patch set 1
Patch Set 1:

Ian: I thought about that, but I thought it might be undesirable to configure RDO if it would not be required, and decided to only configure it if it was clear that openvswitch would need to be installed.  Were it to be configured at a higher level - say, in install_prereqs.sh - either RDO would have to always be enabled for RHEL6 (even if it wouldn't be used) or there would have to be logic extracted for reuse from the quantum setup that could determine whether RDO would be required.

Devstack cores - is the current approach acceptable, or would you like to see one of the alternatives I mentioned?

Patch-set: 1