Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(1 inline comment)

All of the RPMs pulled in here (rdo-release, openvswitch and dependencies) need to be in the appropriate files/rpms/* so they can be loaded in to the pre-built CI environment.  If you do not want them installed with the rest of the deps add '# NOPRIME' to the line so they are only downloaded and then installed by your patch.

Patch-set: 1
Label: Code-Review=-1