Add a new tempest flavor m1.little
Like m1.micro and m1.nano but it has 512 MB of memory.

"little" was chosen arbitrarily because "milli" isn't a colloquial term for
really small like "micro" or "nano"

m1.little will be used by the heat tempest tests to boot full ubuntu or fedora
VMs. m1.tiny is not appropriate because it now enforces a root disk size of 1GB.

Change-Id: I11ea5b79f9b578ef24374c43f80411d2dd6dccf6