Update patch set 2
Patch Set 2:

Aaron, gongysh

gongysh is right. once oslo.config v1.2.0a is used in Quantum, existing conf files using capital letter group names can be used. Thus the conf files are backwards compatible.

The compatibility issue which was considered at that time was conf file only, not code.
The reason was if we also normalized references to group names then 1) too much overhead is added when only quantum had to change the references to lowercase and 2) for minimum code change as possible
(These comments are my views and opinions do not represent those of other oslo cores.)

So I expect it will be a rough ride till we get Quantum to use oslo.config 1.2.0a but it has to be done sooner or later. Lets come up with a plan with minimum confusion.

Patch-set: 2
Label: Code-Review=0