Update patch set 1
Patch Set 1:

given that it's in the tempest configs, and it's for heat, can we just name it m1.small-heat or something. So that people know why it showed up.

It's a nit, but little vs. small is probably going to confuse people playing with devstack.

Dean, feal free to +A over me if you think this is too nit picky.

Patch-set: 1
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0