Update patch set 5
Patch Set 5: I would prefer that you didn't merge this

(3 inline comments)

this seems to go way beyond just configuring the cisco plugin for quantum, as it builds a VM does all kinds of crazy things to it.

Devstack isn't a generic installer, I don't think we should be doing and maintaining all that in here.

Patch-set: 5
Label: Workflow=0