Update patch set 5
Patch Set 5: Do not merge

sorry, I disagree that managing a whole other guest is part of this. It's one thing to point the quantum plugin at an SDN somewhere that someone set up, but using it to bring up the entire SDN is out of bounds for the complexity we want to carry.

Patch-set: 5
Label: Code-Review=-2