Fix url based requires
If an explicit URL is specified inside pip-requires, which results as
the full url in "*.egg-info/dependency_links.txt", devstack passes the
full url as a -f to "pip install". pip seems to require the value of -f
to be a catalog. This change adds an awk string operation, so that only
the url "dirname" will be passed to pip as a find links (-f) parameter.

Change-Id: I4813d2a6a0c99d552238835ad6a96a5050aed400