Fix url based requires
If an explicit URL is specified inside requires, the devstack scripts
used the full url to pass as a -f to pip install. pip requires -f to be
a catalog. This change adds an awk string operation, so that only the
url basename will be passed to pip as a find links parameter.

Change-Id: I4813d2a6a0c99d552238835ad6a96a5050aed400