Set variables in rsyslog.conf to disable rate-limiting.
Per Bug https://bugs.launchpad.net/openstack-ci/+bug/1024487, this will
modify the rate-limiting setting for rsyslog.conf.  If rate-limiting is
being used the addition will set it to 0 otherwise the addition will add
the variables to the configuration file and set them to 0.

Implements: Variables "$SystemLogRateLimitBurst 0", and
"$SystemLogRateLimitInterval 0" in "/etc/rsyslog.conf".

Fixes: Bug 1024487
Change-Id: I2637889cbe9e5e87bbfc0f1bb5047abae34d953c