Update patch set 1
Patch Set 1: (1 inline comment)

The bigger problem is that devstack assumes in all sorts of places that we're using libvirt.

For example, lib/nova uses LIBVIRT_GROUP in many places which do not make sense for a XenServer installation.

Admittedly the "right" fix is to make devstack hypervisor agnostic and not try and install libvirt nomatterwhat but that's a significantly bigger change.

For now we need something that gets XenServer working in devstack again - I'd prefer it was either this change or a revert to https://review.openstack.org/#/c/30703/

Patch-set: 1