Update patch set 1
Patch Set 1:

I'm nervous that "fixing" this could be huge - devstack and specifically nova assume that KVM and libvirt is installed - in fact it's currently a dependency, which is clearly wrong.

"Doing the right thing" and creating appropriate separations could be a large volume of work - and it's certainly currently unmeasurable.

We need a patch that gets XenServer back working again - and actually, I believe that _all_ non-libvirt hypervisors were broken by https://review.openstack.org/#/c/29379/ (including baremetal)

Patch-set: 1