Update patch set 2
Patch Set 2: Do not merge

(1 inline comment)

The packages can not be removed from files/apts/*.  If your goal is to not install them then add # NOPRIME to the end of the lines.  This will delay the installation until they are explicitly called.

There are a number of hard tabs that need to be removed.

I do not see a need for is_libvirt()

Patch-set: 2
Label: Code-Review=-2
Label: Workflow=0