Update patch set 11
Patch Set 11: I would prefer that you didn't merge this

so, one more thing before we merge this. This is buried really deep in the tree so not entirely obvious for people to find. So I feel like there should be a block of comments in either stackrc or lib/quantum explaining the existence of these variables for setting extra options. 

(Somewhere in top level README would be cool as well).

Otherwise, nicely done, I think it will reduce the number of quantum patches that are just to add one config flag.

Patch-set: 11
Label: Code-Review=-1
Label: Workflow=0