Update patch set 11
Patch Set 11:

Thanks Sean, I'll go ahead and add the relevant notes about the usage and resubmit. I see Bob's ML2 devstack patch went in, and I'll have to rebase against that as well. Expect a new version of this patch tomorrow, and thanks for the review!

Patch-set: 11
Label: Code-Review=0