Update patch set 1
Patch Set 1:

Nachi, is the VPNaaS layered on top of Quantum or does it require config changes that can not be done after it starts?  If is just layered I probably prefer it to not be directly in DevStack.  We've already held back on adding other XaaS projects, and this may apply to LBaaS too.  I could be convinced that it needs to be here, I just don't know enough about it and am trying to keep devstack from expanding without bound unnecessarily.

This is why we added extras.d and the ability to drop install/uninstall scripts in there to handle layered stuff.

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0