Update patch set 14
Patch Set 14:

@Sean: I experienced kernel lockups while the guest was writing to the volume, even with the noop scheduler, so the right solution is to skip the loopback device, and use a "real" hard drive. It was not resolved with disabling the secure delete. I ran the old style bfv tests (the one with a separate vm dd-ing bits to a volume), and I had the same lockup issues at that point.

So I had two options:
1. Prepare a stack-volumes before starting devstack (around patchset 11)
2. Modify devstack, so that it could create the stack-volumes group (latest patch)

The "problem" in the xenapi case, is that in order to create the volume group in the early phase of the installation, I had to explicitly install the lvm packages, and the whole solution looked a bit "ugly" (See patchset 11) It hard-coded the name "stack-volumes", and the whole thing was xenapi specific. This was option 1. Option 2, the latter patch makes it possible for other hypervisors to use the same approach, and eliminiated the need for installing lvm early, because at cinder time, the lvm subsystem is already installed.

There are other options as well
- if VOLUME_BACKING_FILE is a block device, skip loopback setup
- use a localrc setting: XEN_CREATE_XVDB, XEN_XVDB_SIZE - to create the XVDB

So on the client side, it would look like:

XEN_CREATE_XVDB=True
VOLUME_BACKING_FILE=/dev/xvdb

What do you think? What is the most complex part of the patch, that you want to see going away (kernel parameters maybe?). Give me some hints, so that I could come up with a better solution.

Mate

Patch-set: 14
Reviewer: Gerrit User 5044 <5044@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0