Update patch set 1
Patch Set 1:

Sean, ok I'll try again to reproduce this issue.

I did a recheck as it works fine in my devstack, the V3 catalog looks ok for me.
It is strange though that a new entry in the catalog affects the keystone V3 token tests. 

I'll dig further into this.

Patch-set: 1