Update patch set 1
Patch Set 1:

It seems that the issue is related to this:
https://blueprints.launchpad.net/keystone/+spec/catalog-optional

As I understood the issue, the length of the catalogue has an impact on the length on the token. When the token is too long,  we get back the following error:

HTTP/1.0 400 Header Line Too Long
Connection: close
Content-length: 0

I wasn't running swift in my devstack, which may be which the test was successful. Any idea how to solve this?

Patch-set: 1