selinux sanity checks
Some checks for selinux to help find some common problems

If selinux is in enforcing mode, then output a warning.  This is
useful for diagnosing possible problems from the logs.

Check for probable installation in homedir and the boolean
httpd_read_user_content flag; if it is not set, then various commands
will hang as requests such as "nova x509-get-root-cert" are routed to
httpd which then can't access necessary files.

Change-Id: I6c635228875ca75e98e707cf8687d7e6d9f02a89