Update patch set 2
Patch Set 2:

Nachi, we ship a extras.d/README with a short description and extras.d/80-tempest.sh as an example.  They really are that simple.  One item that is missing from README is that those scripts are sourced by stack.sh (and unstack.sh) so all config/environment variables are available.

I've got a note to add this to devsyack.org.

Patch-set: 2
Label: Code-Review=0
Label: Workflow=0