xenapi: cleanup VM Installation
VM installation created a machine with multiple network interfaces. This
is a needless complexity at that point. This change removes the network
setup from install-os-vpx.sh script, so that only one network interface
is created, which is connected to the given network. This change also
removes a lot of non-used code from install-os-vpx.sh

Change-Id: Ib0a0a316e849f2c2651305ea657c84820cd2bb3b