Update patch set 3
Patch Set 3:

We're running install_os_domU in dom0 so we should get the IP from there and either inject it into the localrc or, more preferably, xenstore.

Perhaps - since a hosts IP address could change - we should be getting it on host boot and punshing it into xenstore / the VM's xenstore-key.

Patch-set: 3
Label: Code-Review=0