Update patch set 3
Patch Set 3: (1 inline comment)

Okay, so I tried out the xenstore communication, but it seems, xenstore could not be read before the xenfs is mounted, so it might not be available by the time devstack needs it. I would prefer not to modify localrc, it's already a placeholder for too many things, making it an r/w access general purpose communitation channel would not help in this situation. BobBall has strong feelings against using kernel parameters, and also keen to get back the HIMN to play.

So for now, I will remove the HIMN on eth0, and die if I don't have a xenapi url.

As soon as we know how to use HIMN to communicate with xenapi, we could start using that. But for now, let's stay only on the management network.

@John: If we had a HIMN (Host Internal Management Network), and xapi would also have an IP on the OS Management Network, would that work (after making xapi calls to discover the xapi IP on OS Management net)?

Patch-set: 3
Label: Code-Review=0