Use a dummy rpm for Fedora
On Fedora/RHEL python is setup with paths such that pip will install
into /usr rather than /usr/local (see sysconfig.get_paths()).  This
means rpm-packaged and pip-installed packages can't co-exist.

Since we want a later setuptools and pip installed, there is a
chicken-and-egg egg problem: we can't remove the python-setuptools rpm
because it flows through as a dependency to other packages we
need/want.  We can't pip install a later verison, because that will
imperfectly overwrite the packaged files and makes a mess of combined
versions that fails.

The solution presented here is to create a dummy rpm which gets
installed and provides python-setuptools, python-pip and some other
packages that we wish to install via pip, but are also requirements
for other rpm-packages we want (crypto and lxml currently).  The dummy
rpm ensures the dependencies are met and kicks out the existing
packages if installed (via rpm, so no trace is left).  We use
install_pip.sh written by Dean Troyer in [1] to install upstream pip
directly after that.

As a follow-on, we should probably remove all python-* rpm packages
from the Fedora install and let pip do the work.  The only issue is if
other rpm-packages we wish to use bring in python-* packages ... in
that case we probably need the dummy-package work-around for them.

[1] https://review.openstack.org/#/c/39827/

Change-Id: I3ec4ac01c2ec1fb1ce9844f12611d1f86d631991