Install diskimage-builder as a stand-alone service.
If the dib service is enabled (which it is by default),
diskimage-builder and tripleo-image-elements are cloned.

Baremetal has been modified to use the installed diskimage-builder
instead of cloning dib itself.

The bash function disk_image_create is provided for building images,
for example to build a heat tempest test image.

Change-Id: I8a48fead8c2912a276001f3eb03a318f744d7f5f