Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

DiB will need to be added to devstack-gate first in order for devstack to make use of it in the gate (that's why this is currently failing).

We will need to recognize DiB as an official gating project first, and move it to openstack or openstack-dev.

How do you plan to cache the images?

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0