Update patch set 2
Patch Set 2:

Hi Dean,

If you look at unstack.sh, is_service_enabled call is made in many places in a small script. 

I understand your concern that nova service might not be there when unstack.sh is invoked. For example, the user could manually shut it down or it might have just crashed, etc.. However, in a normal circumstances, the change does work at least in the following scenario:
      1. invoke stack.sh 
      2. create some instances
      3. invoke unstack.sh

The above will work in both single-node or multi-node devstack, and it will address the concern described in the bug. 

Please do let me know what your recommendations are to proceed on this.

Patch-set: 2