Update patch set 1
Patch Set 1: Abandoned

Okay, so abandoning it.

It seems, that we haven't arrived yet. If image_id is not specified, nova will not use any image metadata, thus in the xenapi case, the PV-ness or HVM-ness of the image will be guessed by running pygrub in domU (which fails, see bug 1196570), so I end up with an instanve in HVM mode, which works.

But, if image_id is specified, it will launch a PV instance, and that will fail due to a silly bug (something like http://lists.xen.org/archives/html/xen-devel/2013-01/msg00847.html), and that will FAIL.

So I think the real issue in my case is the xen bug.

Patch-set: 1