Update patch set 3
Patch Set 3:

I wanted to keep the same configuration as done on swift v1 since those are used (and make it easier to differentiate between the two what's working or not). If that's really a problem then i can change it but it just would make more modifications on the default config shipped with the samples.

Patch-set: 3
Label: Code-Review=0