Update patch set 2
Patch Set 2: Do not merge

We have been down this path multiple times before and the conclusion has always been to leave selinux in enforcing mode.  The integration work to make (horizon in this case) functional under selinux needs to be done some time, and while DevStack is not necessarily the right place to solve packaging problems it is the right place to work those out if they happen to have impact on the projects directly.

This is a significant change and I do not want to proceed with it on an assumption without some discussion, particularly with those whom it will impact.  Maybe bring it up on the -dev ML (to catch those Fedora/RHEL devs who may not be watching the DevStack commit queue) if you feel strongly that this is the right thing to do?

Patch-set: 2
Label: Code-Review=-2