Modify startup order of Cinder services.
There are cases where the timing between the start up
for cinder-volume and cinder-scheduler service can result
in a race where the scheduler doesn't know about the volume-service
until the next periodic update.

This change does attempts to do an easy fix by swapping the start
order of the cinder services to ensure that the scheduler will be
able to receive the volume service capabilities update.

Fixes bug: 1189595

Change-Id: I8f477ddc04c15c04493f7ce6863e08e1de8f0128