Update patch set 1
Patch Set 1: (1 inline comment)

Sean, I'd actually prefer to have these checks in stack.sh to help show the service dependencies there.  The create_swift_accounts() call could be moved down to where init_swift() is called too, with a check for keystone.

I have been thinking we should group more of the service stuff together, similar to a grenade-style of startup where, for example, all the swift bits are together and its dependencies are all there, as in keystone in this instance.

Patch-set: 1