Update patch set 4
Patch Set 4:

> I don't mind the tmpdir hanging around so much if it is cleaned up in unstack.sh.

That's not so easy to do 100% correctly; what if $TMPDIR was set differently for the stack.sh?  I can just add a rm -rf /tmp/pip-build-* in unstack.sh if you want.  

But I don't think this change makes it any different to the current situation where /tmp/$USER-pip-build is left around anyway?

Patch-set: 4