Update patch set 4
Patch Set 4:

@sdague: This patch actually adds the delete.

The pip install rarely fails (connection issue ?),
but this patch also prevents the future dirty builds.

The unconditional delete will not work properly because the 'set -e' normally cancels the operation, because of the failed build.

I do not see why this fix needs to be more complicated.

Patch-set: 4