Upgrade and remove and upgrade setuptools
The current state of setuptools upgrading is a giant giant giant mess
perpetrated on the masses by the lord of evil. Conflicting versions of
things can wind up on the system and then everything breaks.

To fix this, first install setuptools and pip via pip, then remove the
versions of setuptools and pip that ubuntu installed. Then re-install
pip and setuptools and virtualenv via pip just to be sure.

Change-Id: Iefd9cf90d1459aee92eef19b0328e9e77eaea987