Update patch set 2
Patch Set 2:

it definitely exists for other packages as well. Maybe if we had a # dev comment in line we could take in the -devel packages and just know the difference.

People use devstack for living documentation on what's working, so I think we should be explicit on runtime vs. development deps.

Patch-set: 2