Update projects to global requirements before setup
We've agreed that we should be using global requirements, so
force projects to be updated to global requirements before we
test them.

Also, upgrade and remove and upgrade setuptools

The current state of setuptools upgrading is a giant giant giant mess
perpetrated on the masses by the lord of evil. Conflicting versions of
things can wind up on the system and then everything breaks.

To fix this, first install setuptools and pip via pip, then remove the
versions of setuptools and pip that ubuntu installed. Then re-install
pip and setuptools and virtualenv via pip just to be sure.

Co-Authored-With: Monty Taylor <mordred@inaugust.com>

Change-Id: I0652f639673e600fd7508a9869ec85f8d5ce4518