Update patch set 1
Patch Set 1:

I agree - I guess that production installs are fine because they just use the distro's packages and don't need to track lxml or crypto more closely like we do in devstack for gating.

EPEL doesn't appear to have a newer version of python-lxml (so if the cinder/nova requirements really are for >=2.3 then the 2.2 isn't high enough) - but I assumed that the primary reason we're removing it is so we can rely on updating the requirements.txt files ourselves and automatically updating?

Patch-set: 1