Unfubar "unfubar_setuptools"
Change 0392a10a635a8befb13ff242e2ed5d5be4b23560 apparently introduced
this as "Unbreak the giant mess that is the current state of
setuptools" without explaination of what that giant mess actually is.

get_pip_command assumes pip is called "pip-python" on Fedora based
distros because we are using the packaged pip.  Thus after this code
re-installs upstream pip, which is not renamed, further progress
halts.

I doubt this is the correct solution, but without a better
explaination of the reasoning behind the change it's hard to
understand what is trying to be achieved.

Change-Id: I0b85b2dde1bf4e1f2de71a5c1e7cbd34b6b06e27