Update patch set 1
Patch Set 1:

We've agreed that this probably isn't the right fix - but I was leaving this open to get more input on the right way.

I know that removing python-lxml forceably removes Xen - but this is a generic issue and devstack shouldn't be removing packages and potentially breaking someone's system by also removing the dependencies.

Patch-set: 1