Stop installing PyPI things via apt
It may, and in fact, it is, completely baffling to behold, but it turns
out that the interaction between distro installed and pip installed
libraries is not a good one. In fact, it's a completely and total
disaster which shows clearly the work of a bunch of provicial-minded
people who don't know how to cooperate or admit that someone else might
have a different opinion.

Nonetheless, thus is the world in which we actually exist, theory of how
things should be notwithstanding. We were requested by the distros to
stop managing our own apt repos, so we install our depends via pip now.
Additionally, we now actually even GATE on the pip requirements list.
That means that the requirements list that indicates packages from PyPI
is the absolute and total list of requirements in the form that we can
manage.

Thus - installing those packages also via apt is a not only a waste of
time, it's dangerous because of the undefined interactions between
distro and pip.

Also, we an properly install the dev libs we need here, so that other
dev and test works right.

Closes-Bug: 1203723
Change-Id: I18e480a3312078bde26fc590d4d793e0795e7078