Update patch set 3
Patch Set 3:

An idea that devstack might be able to take from what anvil is doing.

Instead of ignoring the distro's packages completely anvil has taken the approach
that it will scan the distro repo and attempt to use the distros packages
(rpm repos in anvils case - at the current time) only if the versions will
correctly work with the provided pypi requirements.

The main program that helps do this validation that might just be able to be
duplicated or used in devstack (with a corresponding apt compatible version
also)

https://github.com/stackforge/anvil/blob/master/tools/yumfind

  $ ./tools/yumfind  -p 'python-setuptools,setuptools>0.8,<1'
  python-setuptools-0.9.8-0.el6.noarch

For pypi packages that can not be found (or have incompat. version matches) then
will anvil go ahead and attempt to download said missing requirements and build
those into rpms (this build into rpm devstack could avoid since it can just
install them directly).

The other program that helps in this process is the multipip program which anvil
is using to collapse a large set of requirement files into a single set of packages
and versions (which then goes into the finding/downloading/building functions). It
is likely useful also to determine pypi requirement incompatibilities (across git repos).

  $ ./tools/multipip -r glance/test-requirements.txt -r glance/requirements.txt -r nova/test-requirements.txt -r nova/requirements.txt
  Babel>=0.9.6
  Cheetah>=2.4.4
  PasteDeploy>=1.5.0
  SQLAlchemy>=0.7.8,<0.7.99
  WebOb==1.2.3
  amqplib>=0.6.1
  anyjson>=0.2.4
  argparse
  boto
  d2to1>=0.2.10,<0.3
  eventlet>=0.9.17
  greenlet>=0.3.1
  iso8601>=0.1.4
  kombu>=1.0.4
  lxml>=2.3
  netaddr>=0.7.6
  oslo.config>=1.2.0a3
  paramiko
  paste
  pbr>=0.5.16,<0.6
  pyasn1
  pyparsing>=1.5.7,<2.0
  python-cinderclient>=1.0.1
  python-glanceclient>=0.9.0
  python-keystoneclient>=0.2.0
  python-neutronclient>=2.2.3,<3.0.0
  routes>=1.12.3
  six
  sqlalchemy-migrate>=0.7.2
  stevedore>=0.10
  suds>=0.4
  websockify<0.4

Both programs (or derivations of) might be useful in the struggle against package
management systems that I think we all agree are not easy to always work with.

Patch-set: 3
Label: Code-Review=0