Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

so, we've seriously been doing every effort to go in the other direction so that people can mostly get their distro versions, and only have to take the pip versions if their distro is out of date. If we are stripping out all of the packaging versions we might as well be running the gate on arch linux.

I think that if we are going down this path what we actually need is a tool that will act as a go between and install the package if sufficient to pip requirements, and not install it if it won't. Given that we're going to global requirements, sufficiency should be able to be figured out up front.

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0