Update patch set 4
Patch Set 4:

Something like yumfind that Josh mentions and/or building simple packages on the fly is something I'd be ok with - although I'm honestly not sure what the benefit is. We have a global requirements list. It is the source of record for what we do. We are not producing a linux distro. Whether our stuff works with the distro packages of the prior release of a linux distro is honestly not our job. Now - I _do_ think it's still important to test on ubuntu precise and not on arch or something, because there are many other things not encompased in our requirements list that affect things. But if we're talking about python libraries, then either we need to use pip to install the complete set (which is the current case) or we DON'T allow that, and we only allow being compatible with things already in distro so that we can be totally on packges. The second is completely un-viable.

I know we've been pushing more towards packages things, and I'm sorry for my continual part in that. I was wrong. It's mainly due to being a distro supporter for so long, and a person fan of packages. The thing is - we have to solve all of the engineering problems that exist in the world of not using packages, so having solved them, the only think packages for our python dependencies give us is more complexity.

Patch-set: 4
Label: Code-Review=0