Update patch set 4
Patch Set 4:

I agree with the sentiment that its either all pips or all packages and not really anywhere in the middle. Its not really worth the pain to figure out the interactions between pip/python packaging and the distro packaging (apt/yum). But knowing from experience that both routes are painful, u just have to decide which one and stick with it I think in the end. Anvil was initially like devstack and tried to mix them, but that ended about 4 months ago when it was to much a PITA.

Patch-set: 4